+++ test/upgrade_generator_test.rb
@@ -38,7 +38,7 @@ class UpgradeGeneratorTest < Rails::Generators::TestCase
       run_generator
 
       assert_migration("db/migrate/background_schema_migrations_change_unique_index.rb") do |content|
-        assert_includes content, "add_index :background_schema_migrations, [:migration_name, :shard, :connection_class_name]"
+        assert_includes content, "add_index :background_schema_migrations, [:migration_name, :table_name, :shard, :connection_class_name]"
       end
     end
   end